Trying to work on #427

+++ src/gui/image_display.c
@@ -43,7 +43,7 @@ static display_mode last_mode[MAXGRAYVPORT];
 
 /* STF (auto-stretch) data */
 static gboolean stfComputed;	// Flag to know if STF parameters are available
-static double stfShadows, stfHighlights, stfM;
+static float stfShadows, stfHighlights, stfM;
 
 void initialize_image_display() {
 	int i;
